chore: restrict usage of proptest in `noirc_abi`
