+++ tooling/noirc_abi/Cargo.toml
@@ -18,13 +18,13 @@ serde.workspace = true
 thiserror.workspace = true
 num-bigint = "0.4"
 num-traits = "0.2"
-proptest.workspace = true
-proptest-derive.workspace = true
+
 
 [dev-dependencies]
 strum = "0.24"
 strum_macros = "0.24"
-
+proptest.workspace = true
+proptest-derive.workspace = true
 
 [features]
 bn254 = ["acvm/bn254"]
