+++ tooling/noirc_abi/src/arbitrary.rs
@@ -2,8 +2,12 @@ use prop::collection::vec;
 use proptest::prelude::*;
 
 use acvm::{AcirField, FieldElement};
+use iter_extended::{btree_map, vecmap};
 
-use crate::{input_parser::InputValue, Abi, AbiType, InputMap, Sign};
+use crate::{
+    input_parser::InputValue, Abi, AbiParameter, AbiReturnType, AbiType, AbiVisibility, InputMap,
+    Sign,
+};
 use std::collections::{BTreeMap, HashSet};
 
 pub(super) use proptest_derive::Arbitrary;
@@ -20,7 +24,7 @@ pub(super) fn ensure_unique_strings<'a>(iter: impl Iterator<Item = &'a mut Strin
 }
 
 proptest::prop_compose! {
-    pub(super) fn arb_field_from_integer(bit_size: u32)(value: u128)-> FieldElement {
+    fn arb_field_from_integer(bit_size: u32)(value: u128)-> FieldElement {
         let width = (bit_size % 128).clamp(1, 127);
         let max_value = 2u128.pow(width) - 1;
         let value = value % max_value;
@@ -28,7 +32,56 @@ proptest::prop_compose! {
     }
 }
 
-pub(super) fn arb_value_from_abi_type(abi_type: &AbiType) -> SBoxedStrategy<InputValue> {
+fn arb_primitive_abi_type() -> SBoxedStrategy<AbiType> {
+    const MAX_STRING_LEN: u32 = 1000;
+    proptest::prop_oneof![
+        Just(AbiType::Field),
+        Just(AbiType::Boolean),
+        any::<(Sign, u32)>().prop_map(|(sign, width)| {
+            let width = (width % 128).clamp(1, 127);
+            AbiType::Integer { sign, width }
+        }),
+        // restrict length of strings to avoid running out of memory
+        (1..MAX_STRING_LEN).prop_map(|length| AbiType::String { length }),
+    ]
+    .sboxed()
+}
+
+fn arb_abi_type() -> BoxedStrategy<AbiType> {
+    let leaf = arb_primitive_abi_type();
+
+    leaf.prop_recursive(
+        8,   // up to 8 levels deep
+        256, // Shoot for maximum size of 256 nodes
+        10,  // We put up to 10 items per collection
+        |inner| {
+            prop_oneof![
+                (1..10u32, inner.clone())
+                    .prop_map(|(length, typ)| { AbiType::Array { length, typ: Box::new(typ) } })
+                    .boxed(),
+                prop::collection::vec(inner.clone(), 1..10)
+                    .prop_map(|fields| { AbiType::Tuple { fields } })
+                    .boxed(),
+                (".*", prop::collection::vec((".+", inner), 1..10))
+                    .prop_map(|(path, mut fields)| {
+                        // Require that all field names are unique.
+                        ensure_unique_strings(fields.iter_mut().map(|(field_name, _)| field_name));
+                        AbiType::Struct { path, fields }
+                    })
+                    .boxed(),
+            ]
+        },
+    )
+    .boxed()
+}
+
+fn arb_abi_param(typ: AbiType) -> SBoxedStrategy<AbiParameter> {
+    (".+", any::<AbiVisibility>())
+        .prop_map(move |(name, visibility)| AbiParameter { name, typ: typ.clone(), visibility })
+        .sboxed()
+}
+
+fn arb_value_from_abi_type(abi_type: &AbiType) -> SBoxedStrategy<InputValue> {
     match abi_type {
         AbiType::Field => vec(any::<u8>(), 32)
             .prop_map(|bytes| InputValue::Field(FieldElement::from_be_bytes_reduce(&bytes)))
@@ -77,60 +130,34 @@ pub(super) fn arb_value_from_abi_type(abi_type: &AbiType) -> SBoxedStrategy<Inpu
     }
 }
 
-fn arb_primitive_abi_type() -> SBoxedStrategy<AbiType> {
-    const MAX_STRING_LEN: u32 = 1000;
-    proptest::prop_oneof![
-        Just(AbiType::Field),
-        Just(AbiType::Boolean),
-        any::<(Sign, u32)>().prop_map(|(sign, width)| {
-            let width = (width % 128).clamp(1, 127);
-            AbiType::Integer { sign, width }
-        }),
-        // restrict length of strings to avoid running out of memory
-        (1..MAX_STRING_LEN).prop_map(|length| AbiType::String { length }),
-    ]
-    .sboxed()
-}
-
-pub(super) fn arb_abi_type() -> BoxedStrategy<AbiType> {
-    let leaf = arb_primitive_abi_type();
-
-    leaf.prop_recursive(
-        8,   // up to 8 levels deep
-        256, // Shoot for maximum size of 256 nodes
-        10,  // We put up to 10 items per collection
-        |inner| {
-            prop_oneof![
-                (1..10u32, inner.clone())
-                    .prop_map(|(length, typ)| { AbiType::Array { length, typ: Box::new(typ) } })
-                    .boxed(),
-                prop::collection::vec(inner.clone(), 1..10)
-                    .prop_map(|fields| { AbiType::Tuple { fields } })
-                    .boxed(),
-                (".*", prop::collection::vec((".+", inner), 1..10))
-                    .prop_map(|(path, mut fields)| {
-                        // Require that all field names are unique.
-                        ensure_unique_strings(fields.iter_mut().map(|(field_name, _)| field_name));
-                        AbiType::Struct { path, fields }
-                    })
-                    .boxed(),
-            ]
-        },
-    )
-    .boxed()
-}
-
-pub fn arb_input_map(abi: &Abi) -> BoxedStrategy<InputMap> {
-    let values: Vec<_> = abi
-        .parameters
-        .iter()
-        .map(|param| (Just(param.name.clone()), arb_value_from_abi_type(&param.typ)))
-        .collect();
-
-    values
-        .prop_map(|values| {
-            let input_map: InputMap = values.into_iter().collect();
-            input_map
+fn arb_abi_param_and_value() -> BoxedStrategy<(AbiParameter, InputValue)> {
+    arb_abi_type()
+        .prop_flat_map(|typ| {
+            let value = arb_value_from_abi_type(&typ);
+            let param = arb_abi_param(typ);
+            (param, value)
         })
         .boxed()
+}
+
+prop_compose! {
+    fn arb_abi_return_type()
+        (abi_type in arb_abi_type(), visibility: AbiVisibility)
+        -> AbiReturnType {
+            AbiReturnType { abi_type, visibility }
+    }
+}
+
+prop_compose! {
+    pub(super) fn arb_abi_and_input_map()
+        (mut parameters_with_values in proptest::collection::vec(arb_abi_param_and_value(), 0..100), return_type in proptest::option::of(arb_abi_return_type()))
+        -> (Abi, InputMap) {
+            // Require that all parameter names are unique.
+            ensure_unique_strings(parameters_with_values.iter_mut().map(|(param_name,_)| &mut param_name.name));
+
+            let parameters  = vecmap(&parameters_with_values, |(param, _)| param.clone());
+            let input_map = btree_map(parameters_with_values, |(param, value)| (param.name, value));
+
+            (Abi { parameters, return_type, error_types: BTreeMap::default() }, input_map)
+    }
 }
